// apps/main/src/models/v2/dataplane/activeResourceNode.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { ActiveResourceNodeEdges } from './activeResourceNodeEdges';
import type { ActiveResourceNodeMetadata } from './activeResourceNodeMetadata';

/**
 * A node in the active resource graph
 */
export interface ActiveResourceNode {
  /** The deterministic hash of the node in the graph. This hash is also used in the edges map for each node. */
  id: string;
  /** The Project id of the environment of this node. */
  project_id: string;
  /** The environment id. */
  env_id: string;
  resource_type: string;
  resource_class: string;
  resource_id: string;
  deployment_id: string;
  module_id: string;
  module_version: string;
  edges: ActiveResourceNodeEdges;
  metadata: ActiveResourceNodeMetadata;
}